project_root/requirements.txt
# requirements.txt
selenium==4.36.0
pytest==8.4.2
pytest-html==4.1.1
pytest-xdist==3.8.0
webdriver-manager==4.0.2
allure-pytest==2.15.0